+++ tests/test_no_entries.py
@@ -1,0 +1,40 @@
+import sys
+
+from simpleeval import (
+    FunctionNotDefined,
+    NameNotDefined,
+    OperatorNotDefined,
+    SimpleEval,
+)
+
+from .base import DRYTest
+
+
+class TestNoEntries(DRYTest):
+    def test_no_functions(self):
+        self.s.eval("int(42)")
+        with self.assertRaises(FunctionNotDefined):
+            s = SimpleEval(functions={})
+            s.eval("int(42)")
+
+    def test_no_names(self):
+        # does not work on current Py3, True et al. are keywords now
+        self.s.eval("True")
+        # with self.assertRaises(NameNotDefined):
+        s = SimpleEval(names={})
+        if sys.version_info < (3,):
+            with self.assertRaises(NameNotDefined):
+                s.eval("True")
+        else:
+            s.eval("True")
+
+    def test_no_operators(self):
+        self.s.eval("1+2")
+        self.s.eval("~2")
+        s = SimpleEval(operators={})
+
+        with self.assertRaises(OperatorNotDefined):
+            s.eval("1+2")
+
+        with self.assertRaises(OperatorNotDefined):
+            s.eval("~ 2")
