+++ tests/test_operators.py
@@ -1,0 +1,8 @@
+from .base import DRYTest
+
+
+class TestOperators(DRYTest):
+    """Test adding in new operators, removing them, make sure it works."""
+
+    # TODO
+    pass
